+++ CHANGELOG.md
@@ -21,6 +21,13 @@ Upload: git push origin --tags
 Delete tag: git tag -d <tag_name>
 
 ## [Unreleased]
+
+### Changed
+- Non ajax HTTPException now also show error code and message.
+
+### Fixed
+- Fixed missing closing p tag in error html
+
 ## [1.30.2] - 2019-xx-xx
 Cleaned up templates and made small internal adjustments